refactor: api 응답에 맞도록 response type 및 컴포넌트 변경
