+++ src/pages/Ticker/index.tsx
@@ -6,12 +6,12 @@ export default function TickerPage() {
   const { data: TickerData } = useGetPopularTickers();
   return (
     <Wrapper>
-      <TickerList items={TickerData?.content.predictionList} />
+      <TickerList items={TickerData?.content.predictionList || []} />
     </Wrapper>
   );
 }
 const Wrapper = styled.div`
-  width: 100%;
+  width: 90%;
   display: flex;
   flex-direction: column;
   padding: 16px 0;
